Pull Request

+++ editortools/filter.py
@@ -342,6 +342,17 @@ class FilterToolPanel(Panel):
 
         tool = self.tool
 
+        for (i, _) in config.config._sections["Filter Keys"].items():
+            if i == "__name__":
+                continue;
+            There = False
+            for j in tool.filterModules:
+                if i == j.lower():
+                    There = True
+                    break;
+            if not There:
+                config.config.remove_option("Filter Keys", i)
+
         if len(tool.filterModules) is 0:
             self.add(Label("No filter modules found!"))
             self.shrink_wrap()
@@ -743,6 +754,7 @@ class FilterTool(EditorTool):
                 for lib in filt.libraries:
                     lib_path = os.path.join(directories.getFiltersDir(), 'lib', lib["path"].replace("<sep>", os.path.sep))
                     lib_path_stock = os.path.join(directories.getDataDir(), 'stock-filters', 'lib', lib["path"].replace("<sep>", os.path.sep))
+                    print lib_path, lib_path_stock
                     if os.path.exists(lib_path):
                         sys.path.append(lib_path)
                     elif os.path.exists(lib_path_stock):
@@ -816,6 +828,8 @@ class FilterTool(EditorTool):
         def tryImport(name):
             module_file_object = open(os.path.join(directories.getFiltersDir(), name))
             module_name = name.split(os.path.sep)[-1].replace(".py", "")
+            sys.path.append(os.path.join(directories.getFiltersDir(), 'lib', "Bundled Libraries"))
+            sys.path.append(os.path.join(directories.getDataDir(), 'stock-filters', 'lib', "Bundled Libraries"))
             try:
                 m = imp.load_module(module_name, module_file_object, name, ('.py', 'rb', imp.PY_SOURCE))
                 #!# Fix for Naor's filter bug
@@ -877,7 +891,7 @@ class FilterTool(EditorTool):
                 #!# Fix for Naor's filter bug
 #                if os.sys.platform == 'darwin' or True: # for debug
                 if os.sys.platform == 'darwin':
-                    if os.path.exists(m.__file__):
+                    if os.path.exists(m.__file__) and m.__file__.endswith('.pyc'):
                         os.remove(m.__file__)
                 #!#
             except Exception, e:
